move unit tests to separate folder
