+++ tests/unit/Api/CustomerApiTest.php
@@ -1,9 +1,11 @@
 <?php
 declare(strict_types=1);
+/**
+ */
 
-namespace CommerceLeague\ActiveCampaignApi\tests\Api;
+namespace CommerceLeague\ActiveCampaignApi\tests\unit\Api;
 
-use CommerceLeague\ActiveCampaignApi\Api\ContactApi;
+use CommerceLeague\ActiveCampaignApi\Api\CustomerApi;
 use CommerceLeague\ActiveCampaignApi\Client\CommonResourceClientInterface;
 use CommerceLeague\ActiveCampaignApi\Paginator\Page;
 use CommerceLeague\ActiveCampaignApi\Paginator\PageFactoryInterface;
@@ -12,9 +14,9 @@ use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
 /**
- * Class ContactApiTest
+ * Class CustomerApiTest
  */
-class ContactApiTest extends TestCase
+class CustomerApiTest extends TestCase
 {
     /**
      * @var MockObject|CommonResourceClientInterface
@@ -32,16 +34,16 @@ class ContactApiTest extends TestCase
     protected $cursorFactory;
 
     /**
-     * @var ContactApi
+     * @var CustomerApi
      */
-    protected $contactApi;
+    protected $customerApi;
 
     protected function setUp()
     {
         $this->resourceClient = $this->createMock(CommonResourceClientInterface::class);
         $this->pageFactory = $this->createMock(PageFactoryInterface::class);
         $this->cursorFactory = $this->createMock(ResourceCursorFactoryInterface::class);
-        $this->contactApi = new ContactApi(
+        $this->customerApi = new CustomerApi(
             $this->resourceClient,
             $this->pageFactory,
             $this->cursorFactory
@@ -54,10 +56,10 @@ class ContactApiTest extends TestCase
         $response = ['response'];
         $this->resourceClient->expects($this->once())
             ->method('getResource')
-            ->with('api/3/contacts/%s', [$id])
+            ->with('api/3/ecomCustomers/%s', [$id])
             ->willReturn($response);
 
-        $this->assertEquals($response, $this->contactApi->get($id));
+        $this->assertEquals($response, $this->customerApi->get($id));
     }
 
     public function testListPerPage()
@@ -67,9 +69,9 @@ class ContactApiTest extends TestCase
         $queryParameters = ['query' => 'param'];
 
         $response = [
-            'contacts' => [
-                ['first contact'],
-                ['second contact']
+            'ecomCustomers' => [
+                ['first ecomCustomer'],
+                ['second ecomCustomer']
             ],
             'meta' => [
                 'total' => 1000
@@ -79,7 +81,7 @@ class ContactApiTest extends TestCase
         $this->resourceClient->expects($this->once())
             ->method('getResources')
             ->with(
-                'api/3/contacts',
+                'api/3/ecomCustomers',
                 [],
                 $limit,
                 $offset,
@@ -89,29 +91,31 @@ class ContactApiTest extends TestCase
 
         $this->pageFactory->expects($this->once())
             ->method('createPage')
-            ->with($this->contactApi, $response['contacts'], $response['meta']);
+            ->with($this->customerApi, $response['ecomCustomers'], $response['meta']);
 
-        $this->contactApi->listPerPage($limit, $offset, $queryParameters);
+        $this->customerApi->listPerPage($limit, $offset, $queryParameters);
     }
 
     public function testAll()
     {
         $limit = 55;
         $queryParameters = ['query' => 'param'];
+
         $response = [
-            'contacts' => [
-                ['first contact'],
-                ['second contact']
+            'ecomCustomers' => [
+                ['first ecomCustomer'],
+                ['second ecomCustomer']
             ],
             'meta' => [
                 'total' => 1000
             ]
         ];
 
+
         $this->resourceClient->expects($this->once())
             ->method('getResources')
             ->with(
-                'api/3/contacts',
+                'api/3/ecomCustomers',
                 [],
                 $limit,
                 0,
@@ -124,14 +128,14 @@ class ContactApiTest extends TestCase
 
         $this->pageFactory->expects($this->once())
             ->method('createPage')
-            ->with($this->contactApi, $response['contacts'], $response['meta'])
+            ->with($this->customerApi, $response['ecomCustomers'], $response['meta'])
             ->willReturn($page);
 
         $this->cursorFactory->expects($this->once())
             ->method('createCursor')
             ->with($limit, $page);
 
-        $this->contactApi->all($limit, $queryParameters);
+        $this->customerApi->all($limit, $queryParameters);
     }
 
     public function testCreate()
@@ -141,10 +145,10 @@ class ContactApiTest extends TestCase
 
         $this->resourceClient->expects($this->once())
             ->method('createResource')
-            ->with('api/3/contacts', [], $data)
+            ->with('api/3/ecomCustomers', [], $data)
             ->willReturn($response);
 
-        $this->assertEquals($response, $this->contactApi->create($data));
+        $this->assertEquals($response, $this->customerApi->create($data));
     }
 
     public function testUpdate()
@@ -155,23 +159,10 @@ class ContactApiTest extends TestCase
 
         $this->resourceClient->expects($this->once())
             ->method('updateResource')
-            ->with('api/3/contacts/%s', [$id], $data)
+            ->with('api/3/ecomCustomers/%s', [$id], $data)
             ->willReturn($response);
 
-        $this->assertEquals($response, $this->contactApi->update($id, $data));
-    }
-
-    public function testUpsert()
-    {
-        $data = ['data'];
-        $response = ['response'];
-
-        $this->resourceClient->expects($this->once())
-            ->method('upsertResource')
-            ->with('api/3/contact/sync', [], $data)
-            ->willReturn($response);
-
-        $this->assertEquals($response, $this->contactApi->upsert($data));
+        $this->assertEquals($response, $this->customerApi->update($id, $data));
     }
 
     public function testDelete()
@@ -181,9 +172,9 @@ class ContactApiTest extends TestCase
 
         $this->resourceClient->expects($this->once())
             ->method('deleteResource')
-            ->with('api/3/contacts/%s', [$id])
+            ->with('api/3/ecomCustomers/%s', [$id])
             ->willReturn($response);
 
-        $this->assertEquals($response, $this->contactApi->delete($id));
+        $this->assertEquals($response, $this->customerApi->delete($id));
     }
 }
