+++ tests/unit/Client/CommonResourceClientTest.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 /**
  */
 
-namespace CommerceLeague\ActiveCampaignApi\tests\Client;
+namespace CommerceLeague\ActiveCampaignApi\tests\unit\Client;
 
 use CommerceLeague\ActiveCampaignApi\Client\CommonResourceClient;
 use CommerceLeague\ActiveCampaignApi\Client\HttpClientInterface;
