+++ tests/unit/Client/HttpClientTest.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 /**
  */
 
-namespace CommerceLeague\ActiveCampaignApi\tests\Client;
+namespace CommerceLeague\ActiveCampaignApi\tests\unit\Client;
 
 use CommerceLeague\ActiveCampaignApi\Client\HttpClient;
 use PHPUnit\Framework\MockObject\MockObject;
