+++ tests/unit/Client/HttpExceptionHandlerTest.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 /**
  */
 
-namespace CommerceLeague\ActiveCampaignApi\tests\Client;
+namespace CommerceLeague\ActiveCampaignApi\tests\unit\Client;
 
 use CommerceLeague\ActiveCampaignApi\Client\HttpExceptionHandler;
 use CommerceLeague\ActiveCampaignApi\Exception\BadRequestHttpException;
